# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "AppObras API",
        "description": "API v1 de autenticação e conta",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API base"
        }
    ],
    "paths": {
...
        "/api/v1/projects/{project}/phases": {
            "get": {
                "tags": [
                    "Phases"
                ],
                "summary": "Listar fases do projeto",
                "description": "Retorna todas as fases de um projeto com progresso calculado",
                "operationId": "d7a760c08e07f540d5304de0d9fb0c8d",
                "parameters": [
                    {
                        "name": "X-Company-Id",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "project",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "draft",
                                "active",
                                "archived"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de fases com progresso"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}